switch pressure to uint8.

+++ internal/desktop/xinput.go
@@ -6,7 +6,7 @@ func (manager *DesktopManagerCtx) inputRelToAbs(x, y int) (int, int) {
 	return (x * xinput.AbsX) / manager.screenSize.Width, (y * xinput.AbsY) / manager.screenSize.Height
 }
 
-func (manager *DesktopManagerCtx) TouchBegin(touchId uint32, x, y int, pressure uint16) error {
+func (manager *DesktopManagerCtx) TouchBegin(touchId uint32, x, y int, pressure uint8) error {
 	mu.Lock()
 	defer mu.Unlock()
 
@@ -14,7 +14,7 @@ func (manager *DesktopManagerCtx) TouchBegin(touchId uint32, x, y int, pressure 
 	return manager.input.TouchBegin(touchId, x, y, pressure)
 }
 
-func (manager *DesktopManagerCtx) TouchUpdate(touchId uint32, x, y int, pressure uint16) error {
+func (manager *DesktopManagerCtx) TouchUpdate(touchId uint32, x, y int, pressure uint8) error {
 	mu.Lock()
 	defer mu.Unlock()
 
@@ -22,7 +22,7 @@ func (manager *DesktopManagerCtx) TouchUpdate(touchId uint32, x, y int, pressure
 	return manager.input.TouchUpdate(touchId, x, y, pressure)
 }
 
-func (manager *DesktopManagerCtx) TouchEnd(touchId uint32, x, y int, pressure uint16) error {
+func (manager *DesktopManagerCtx) TouchEnd(touchId uint32, x, y int, pressure uint8) error {
 	mu.Lock()
 	defer mu.Unlock()
