+++ pkg/xinput/dummy.go
@@ -18,14 +18,14 @@ func (d *dummy) Close() error {
 
 func (d *dummy) Debounce(duration time.Duration) {}
 
-func (d *dummy) TouchBegin(touchId uint32, x, y int, pressure uint16) error {
+func (d *dummy) TouchBegin(touchId uint32, x, y int, pressure uint8) error {
 	return nil
 }
 
-func (d *dummy) TouchUpdate(touchId uint32, x, y int, pressure uint16) error {
+func (d *dummy) TouchUpdate(touchId uint32, x, y int, pressure uint8) error {
 	return nil
 }
 
-func (d *dummy) TouchEnd(touchId uint32, x, y int, pressure uint16) error {
+func (d *dummy) TouchEnd(touchId uint32, x, y int, pressure uint8) error {
 	return nil
 }
